[#15] feature : 현재 시세 조회 기능 추가

+++ src/main/java/zb/accountMangement/stock/domain/Stock.java
@@ -21,5 +21,5 @@ public class Stock {
 
   private String name; // 종목명
 
-  private Long currentPrice ;   //// ??? 현재가 ????
+  private Double currentPrice ;   //// ??? 현재가 ????
 }
